Fix profane phrases with spaces
If a phrase has a space, it must be in the regex section. Spaces don't
work in the non-regex section because of the way the string scanning
words: we split the given strings on spaces for those items.

Another limitation is that parentheses must use the non-capturing `?:`,
otherwise it breaks the `last_match` code.

The code is in need of refactoring at this point because the
translations file is a bit hard to read, including the quirks above.
This fix will keep it limping along until that day.